--- test.cpp
@@ -1,13 +1,0 @@
-#include "clean.h"
-
-#include <vector>
-int main()
-{
-	std::vector<int> x;
-	{
-		using namespace clean;
-
-	}
-
-	return 0;
-}
