+++ tests/test.cpp
@@ -1,0 +1,7 @@
+#include "../clean.h"
+
+
+int main()
+{
+    clean::io::println("hello world");
+}
